modules: network: channel definition into module header
Network channel definition into dedicated network header.
