+++ app/src/common/message_channel.c
@@ -18,14 +18,6 @@ ZBUS_CHAN_DEFINE(TRIGGER_CHAN,
 		 ZBUS_MSG_INIT(0)
 );
 
-ZBUS_CHAN_DEFINE(NETWORK_CHAN,
-		 struct network_msg,
-		 NULL,
-		 NULL,
-		 ZBUS_OBSERVERS_EMPTY,
-		 ZBUS_MSG_INIT(.type = NETWORK_DISCONNECTED)
-);
-
 ZBUS_CHAN_DEFINE(ERROR_CHAN,
 		 enum error_type,
 		 NULL,
