+++ app/src/modules/cloud/cloud_module.c
@@ -20,6 +20,7 @@
 #include "modules_common.h"
 #include "message_channel.h"
 #include "battery.h"
+#include "network.h"
 
 /* Register log module */
 LOG_MODULE_REGISTER(cloud, CONFIG_APP_CLOUD_LOG_LEVEL);
