+++ app/src/modules/led/led.c
@@ -13,6 +13,7 @@
 #include "led_pwm.h"
 #include "led.h"
 #include "location.h"
+#include "network.h"
 
 /* Register log module */
 LOG_MODULE_REGISTER(led, CONFIG_APP_LED_LOG_LEVEL);
