+++ tests/module/cloud/CMakeLists.txt
@@ -23,6 +23,7 @@ zephyr_include_directories(${ZEPHYR_BASE}/subsys/testsuite/include)
 zephyr_include_directories(../../../app/src/modules/cloud)
 zephyr_include_directories(../../../app/src/common)
 zephyr_include_directories(../../../app/src/modules/battery)
+zephyr_include_directories(../../../app/src/modules/network)
 zephyr_include_directories(${NRF_DIR}/subsys/net/lib/nrf_cloud/include)
 zephyr_include_directories(${NRF_DIR}/../modules/lib/cjson)
